Adequando o nome dos atributos e parametros (no dominio e na persistencia)
*Usuario

+++ app/src/main/java/com/apphamba/hamba/usuario/dominio/Usuario.java
@@ -6,11 +6,11 @@ public class Usuario {
     private String email;
     private String senha;
 
-    public int getIdUsuario() {
+    public long getIdUsuario() {
         return idUsuario;
     }
 
-    public void setIdUsuario(int idUsuario) {
+    public void setIdUsuario(long idUsuario) {
         this.idUsuario = idUsuario;
     }
